Switch to input buffering on Android
Input buffering is implicitly used by event accumulation, but this commit makes it more generic so it can be enabled for other uses.

So far it's only enabled for Android (it should be possible for iOS as well; for desktop OSs it's currently not feasible given main and UI threads are the same).

Key, touch and joystick events will be passed directly from the UI thread to Godot, so they can be buffered and processed at any point, which will come in handy for the next commit. As another consequence of this new way of passing events, less Java object are created at runtime (`Runnable`), which is good since the garbage collector needs to run less.

Sensors, go back, etc. will be handled as formerly. Anyway, this commit by iself won't change observed behavior for the others either.